+++ lib/reusable_list_item.dart
@@ -1,5 +1,5 @@
 export 'views/list_items/app_list_item.dart';
-export 'views/list_items/generic_slider_widget.dart';
+export 'views/list_items/app_slider_widget.dart';
 export 'views/list_items/image_with_title_card.dart';
 export 'views/list_items/info_card.dart';
 export 'domain/entity/info_entity.dart';